movie-popularity-predictor/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.33.0
joblib==1.3.2
pandas==2.1.4
scikit-learn==1.3.2
nltk==3.8.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
scipy==1.11.3
apache-airflow-providers-postgres
apache-airflow-providers-common-io